made changes to make different webpack config for local and dev so as to run webpack dev server only on local
