+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build-dev": "webpack --config webpack.dev.js",
-    "dev": "webpack serve --config webpack.dev.js"
+    "local": "webpack serve --config webpack.local.js"
   },
   "author": "",
   "license": "ISC",
