+++ webpack.common.js
@@ -1,0 +1,58 @@
+const path = require("path");
+const HTMLWebpackPlugin = require("html-webpack-plugin");
+const { CleanWebpackPlugin } = require("clean-webpack-plugin");
+
+const buildPath = path.resolve(__dirname, "dist");
+
+const config = {
+  mode: "development",
+  entry: {
+    app: "./src/index.js",
+  },
+  output: {
+    filename: "[name].[contenthash:6].js",
+    path: buildPath,
+    publicPath: "/assets",
+  },
+  module: {
+    // loaders transforms the assets
+    rules: [
+      {
+        test: /\.(js|jsx)$/,
+        exclude: /(node_modules)/,
+        use: {
+          loader: "babel-loader",
+          options: {
+            presets: ["@babel/preset-env", "@babel/preset-react"],
+          },
+        },
+      },
+    ],
+  },
+  optimization: {
+    splitChunks: {
+      chunks: "all",
+      // name: true,
+      cacheGroups: {
+        vendor: {
+          test: /[\\/]node_modules[\\/]/,
+          priority: -10,
+        },
+        default: {
+          minChunks: 2,
+          priority: -20,
+          reuseExistingChunk: true,
+        },
+      },
+    },
+  },
+  plugins: [
+    new CleanWebpackPlugin(),
+    new HTMLWebpackPlugin({
+      template: "template.html",
+      filename: "index.html",
+    }),
+  ],
+};
+
+module.exports = config;
