+++ webpack.dev.js
@@ -1,62 +1,11 @@
-const path = require("path");
-const HTMLWebpackPlugin = require("html-webpack-plugin");
-const { CleanWebpackPlugin } = require("clean-webpack-plugin");
+const commonConfig = require("./webpack.common");
 
-const buildPath = path.resolve(__dirname, "dist");
-
-const config = {
-  mode: "development",
-  entry: {
-    app: "./src/index.js",
-  },
-  output: {
-    filename: "[name].[contenthash:6].js",
-    path: buildPath,
-    publicPath: "/assets",
-  },
-  devServer: {
-    port: 3000,
-  },
+const devConfig = {
+  ...commonConfig,
+  //   devServer: {
+  //     port: 3000,
+  //   },
   devtool: "inline-source-map",
-  module: {
-    // loaders transforms the assets
-    rules: [
-      {
-        test: /\.(js|jsx)$/,
-        exclude: /(node_modules)/,
-        use: {
-          loader: "babel-loader",
-          options: {
-            presets: ["@babel/preset-env", "@babel/preset-react"],
-          },
-        },
-      },
-    ],
-  },
-  optimization: {
-    splitChunks: {
-      chunks: "all",
-      // name: true,
-      cacheGroups: {
-        vendor: {
-          test: /[\\/]node_modules[\\/]/,
-          priority: -10,
-        },
-        default: {
-          minChunks: 2,
-          priority: -20,
-          reuseExistingChunk: true,
-        },
-      },
-    },
-  },
-  plugins: [
-    new CleanWebpackPlugin(),
-    new HTMLWebpackPlugin({
-      template: "template.html",
-      filename: "index.html",
-    }),
-  ],
 };
 
-module.exports = config;
+module.exports = devConfig;
